reconfigured directory
