+++ app/app.py
@@ -3,8 +3,5 @@ from app.db import db
 
 app = create_app()
 
-with app.app_context():
-    db.create_all()  # Create all tables (if not existing)
-
 if __name__ == '__main__':
     app.run(debug=True)
