+++ app/static/search.js
