+++ app/templates/compound_info.html
